chore: move SvgNode to its own class; extract asNode functions to targeted file
